Added Junit testing

+++ src/main/java/org/calculator/Calculator.java
@@ -15,8 +15,8 @@ public class Calculator {
 
         return result;
     }
-    public long power(int base, int exponent){
-        return (long) Math.pow(base, exponent);
+    public double power(double base, double exponent){
+        return Math.pow(base, exponent);
     }
 
     public double naturalLog(double x){
